Make Input handler: InputManager OOP
